Move searchbar widget to seperate file
